+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+2020-03-17 (0.1.0)
+------------------
+- There is now a new class `Variable` in `qualg.scalars` which represents a complex number as a symbolic variable.
+  See quickstart guide for how to use it.
+
 2020-03-17 (0.0.3)
 ------------------
 - Refactoring, fixing small bugs